+++ RELEASE-NOTES.md
@@ -1,3 +1,9 @@
+# February 20, 2026 Release Notes - 0.2.5
+
+## Updates
+1. [Service Connectors module](./service-connectors/)
+  - Enhancement: Added monitoring as source and target in service connectors.
+
 # October 10, 2025 Release Notes - 0.2.4
 
 ## Updates